oinkr-bot/backup/requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==20.3
python-dotenv
gspread
google-auth